added missing documentation
