+++ CodeEdit/Info.plist
@@ -45,6 +45,6 @@
 		</dict>
 	</array>
 	<key>GitHash</key>
-	<string>73db927603933074fe067eead100659caa76d87d</string>
+	<string>0573dd9eda65e78abcd878d6c912e513bea04dbd</string>
 </dict>
 </plist>
